+++ openjpa-jdbc/src/main/java/org/apache/openjpa/jdbc/kernel/exps/ContainsExpression.java
@@ -52,7 +52,7 @@ class ContainsExpression
             if (count == null)
                 count = 0;
             else
-                count = count.intValue() + 1;
+                count = count + 1;
             contains.put(path, count);
 
             sql.setContainsId(count.toString());
