+++ openjpa-lib/src/main/java/org/apache/openjpa/lib/rop/RandomAccessResultList.java
@@ -118,7 +118,7 @@ public class RandomAccessResultList extends AbstractNonSequentialResultList {
     private Object instantiateRow(Integer i) {
         _requests++;
         try {
-            if (!_rop.absolute(i.intValue()))
+            if (!_rop.absolute(i))
                 return PAST_END;
 
             Object ob = _rop.getResultObject();
@@ -157,7 +157,7 @@ public class RandomAccessResultList extends AbstractNonSequentialResultList {
         Integer key;
         for (Iterator itr = _rows.keySet().iterator(); itr.hasNext(); count++) {
             key = (Integer) itr.next();
-            full[key.intValue()] = _rows.get(key);
+            full[key] = _rows.get(key);
         }
 
         // double-check, in case any of the soft references were
