+++ openjpa-persistence-jdbc/src/test/java/org/apache/openjpa/persistence/jdbc/sqlcache/ParentId.java
@@ -79,7 +79,7 @@ public class ParentId implements java.io.Serializable {
             String hashStr = this.id + ":" + this.name + ":" + this.addrId;
             hashcode = hashStr.hashCode();
         }
-        return hashcode.intValue();
+        return hashcode;
     }
 
 }
