+++ openjpa-persistence-jdbc/src/test/java/org/apache/openjpa/persistence/meta/common/apps/SequenceAssigned3.java
@@ -48,7 +48,7 @@ public class SequenceAssigned3 {
 
         @Override
         public int hashCode() {
-            return (int) (pk == null ? 0 : pk.longValue()
+            return (int) (pk == null ? 0 : pk
                 % (long) Integer.MAX_VALUE);
         }
